pyproject.toml
[project]
name = "marathi-meme-generator"
version = "0.1.0"
description = "A Streamlit application that generates memes with Marathi text"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit>=1.32.0",
    "Pillow>=10.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python = "3.9"

[tool.hatch.metadata]
allow-direct-references = true 
